requirements.txt
pandas
numpy
scikit-learn
scipy
matplotlib
seaborn
opencv-python
Pillow
streamlit
tensorflow